Fixed flipped tiles because gid was stored in a int instead of an unsigned, cleared a bit this messy code

+++ Expressions.cpp
@@ -2,57 +2,49 @@
 
 #define __(x) Runtime.CopyString(x)
 
-char *Extension::getLastError()
-{
+char *Extension::getLastError() {
 	if (!mLastError)
 		return __("No errors");
 	return __(mLastError);
 }
 
-int Extension::getMapWidth()
-{
+int Extension::getMapWidth() {
 	if (!mMap)
 		return 0;
 	return mMap->getWidth();
 }
 
-int Extension::getMapHeight()
-{
+int Extension::getMapHeight() {
 	if (!mMap)
 		return 0;
 	return mMap->getHeight();
 }
 
-int Extension::getMapTileWidth()
-{
+int Extension::getMapTileWidth() {
 	if (!mMap)
 		return 0;
 	return mMap->getTileWidth();
 }
 
-int Extension::getMapTileHeight()
-{
+int Extension::getMapTileHeight() {
 	if (!mMap)
 		return 0;
 	return mMap->getTileHeight();
 }
 
-char *Extension::getMapOrientation()
-{
+char *Extension::getMapOrientation() {
 	if (!mMap)
 		return __("");
 	return __(mMap->getOrientation().c_str());
 }
 
-char *Extension::getMapVersion()
-{
+char *Extension::getMapVersion() {
 	if (!mMap)
 		return __("");
 	return __(mMap->getVersion().c_str());
 }
 
-char *Extension::getMapProperty(char *name, char *defaultValue)
-{
+char *Extension::getMapProperty(char *name, char *defaultValue) {
 	if (!mMap)
 		return __(defaultValue);
 	std::map<std::string, std::string> properties = mMap->getProperties();
@@ -63,93 +55,80 @@ char *Extension::getMapProperty(char *name, char *defaultValue)
 	return __(defaultValue);
 }
 
-char *Extension::getTilesetName()
-{
+char *Extension::getTilesetName() {
 	if (!mTileset)
 		return __("");
 	return __(mTileset->getName().c_str());
 }
 
-char *Extension::getTilesetImagePath()
-{
+char *Extension::getTilesetImagePath() {
 	if (!mTileset)
 		return __("");
 	return __(mTileset->getImageSource().c_str());
 }
 
-int Extension::getTilesetId()
-{
+int Extension::getTilesetId() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getId();
 }
 
-int Extension::getTilesetFirstGid()
-{
+int Extension::getTilesetFirstGid() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getFirstGid();
 }
 
-int Extension::getTilesetWidth()
-{
+int Extension::getTilesetWidth() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getWidth();
 }
 
-int Extension::getTilesetHeight()
-{
+int Extension::getTilesetHeight() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getHeight();
 }
 
-int Extension::getTilesetTileWidth()
-{
+int Extension::getTilesetTileWidth() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getTileWidth();
 }
 
-int Extension::getTilesetTileHeight()
-{
+int Extension::getTilesetTileHeight() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getTileHeight();
 }
 
-int Extension::getTilesetSpacing()
-{
+int Extension::getTilesetSpacing() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getSpacing();
 }
 
-int Extension::getTilesetMargin()
-{
+int Extension::getTilesetMargin() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getMargin();
 }
 
-int Extension::getTilesetXTileOffset()
-{
+int Extension::getTilesetXTileOffset() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getOffsetX();
 }
 
-int Extension::getTilesetYTileOffset()
-{
+int Extension::getTilesetYTileOffset() {
 	if (!mTileset)
 		return 0;
 	return mTileset->getOffsetY();
 
 }
 
-char *Extension::getTilesetProperty(char *name, char *defaultValue)
-{
+char *Extension::getTilesetProperty(char *name, char *defaultValue) {
 	if (!mTileset)
 		return __(defaultValue);
 	std::map<std::string, std::string> properties = mTileset->getProperties();
@@ -160,43 +139,37 @@ char *Extension::getTilesetProperty(char *name, char *defaultValue)
 	return __(defaultValue);
 }
 
-char *Extension::getLayerName()
-{
+char *Extension::getLayerName() {
 	if (!mLayer)
 		return 0;
 	return __(mLayer->getName().c_str());
 }
 
-int Extension::getLayerWidth()
-{
+int Extension::getLayerWidth() {
 	if (!mLayer)
 		return 0;
 	return mLayer->getWidth();
 }
 
-int Extension::getLayerHeight()
-{
+int Extension::getLayerHeight() {
 	if (!mLayer)
 		return 0;
 	return mLayer->getHeight();
 }
 
-float Extension::getLayerOpacity()
-{
+float Extension::getLayerOpacity() {
 	if (!mLayer)
 		return 1;
 	return mLayer->getOpacity();
 }
 
-int Extension::isLayerVisible()
-{
+int Extension::isLayerVisible() {
 	if (!mLayer)
 		return 1;
 	return mLayer->getVisible();
 }
 
-char *Extension::getLayerProperty(char *name, char *defaultValue)
-{
+char *Extension::getLayerProperty(char *name, char *defaultValue) {
 	if (!mLayer)
 		return __(defaultValue);
 	std::map<std::string, std::string> properties = mLayer->getProperties();
@@ -207,89 +180,76 @@ char *Extension::getLayerProperty(char *name, char *defaultValue)
 	return __(defaultValue);
 }
 
-int Extension::getTileGid()
-{
+int Extension::getTileGid() {
 	if (!mTile)
 		return 0;
 	return mTile->getGid();
 }
 
-int Extension::getTileWidth()
-{
+int Extension::getTileWidth() {
 	if (!mTile)
 		return 0;
 	return mTile->getWidth();
 }
 
-int Extension::getTileHeight()
-{
+int Extension::getTileHeight() {
 	if (!mTile)
 		return 0;
 	return mTile->getHeight();
 }
 
-int Extension::getTilePositionOnMapX()
-{
+int Extension::getTilePositionOnMapX() {
 	if (!mTile)
 		return 0;
 	return mTile->getX() + mOffsetX;
 }
 
-int Extension::getTilePositionOnMapY()
-{
+int Extension::getTilePositionOnMapY() {
 	if (!mTile)
 		return 0;
 	return mTile->getY() + mOffsetY;
 }
 
-int Extension::getTileAssociatedTilesetId()
-{
+int Extension::getTileAssociatedTilesetId() {
 	if (!mTile)
 		return 0;
 	return mTile->getTilesetId();
 }
 
-int Extension::getTilePositionOnTilesetX()
-{
+int Extension::getTilePositionOnTilesetX() {
 	if (!mTile)
 		return 0;
 	return mTile->getTilesetX();
 }
 
-int Extension::getTilePositionOnTilesetY()
-{
+int Extension::getTilePositionOnTilesetY() {
 	if (!mTile)
 		return 0;
 	return mTile->getTilesetY();
 }
 
-int Extension::isTileFlippedHorizontally()
-{
+int Extension::isTileFlippedHorizontally() {
 	if (!mTile)
 		return 0;
 	return mTile->getFlippedHorizontally();
 }
 
-int Extension::isTileFlippedVertically()
-{
+int Extension::isTileFlippedVertically() {
 	if (!mTile)
 		return 0;
 	return mTile->getFlippedVertically();
 }
 
-int Extension::isTileFlippedDiagonally()
-{
+int Extension::isTileFlippedDiagonally() {
 	if (!mTile)
 		return 0;
 	return mTile->getFlippedDiagonally();
 }
 
-char *Extension::getTileProperty(char *name, char *defaultValue)
-{
+char *Extension::getTileProperty(char *name, char *defaultValue) {
 	if (!mTile)
 		return __(defaultValue);
-	try
-	{
+	try {
 		const auto &tileset = mMap->getTilesets().at(mTile->getTilesetId());
 		if (!tileset)
 			return __(defaultValue);
@@ -300,56 +260,48 @@ char *Extension::getTileProperty(char *name, char *defaultValue)
 			return __(it->second.c_str());
 		return __(defaultValue);
 	}
-	catch (std::out_of_range &)
-	{
+	catch (std::out_of_range&) {
 		return __(defaultValue);
 	}
 }
 
-char *Extension::getObjectGroupName()
-{
+char *Extension::getObjectGroupName() {
 	if (!mObjectGroup)
 		return __("");
 	return __(mObjectGroup->getName().c_str());
 }
 
-char *Extension::getObjectGroupDrawOrder()
-{
+char *Extension::getObjectGroupDrawOrder() {
 	if (!mObjectGroup)
 		return __("");
 	return __(mObjectGroup->getDrawOrder().c_str());
 }
 
-int Extension::getObjectGroupWidth()
-{
+int Extension::getObjectGroupWidth() {
 	if (!mObjectGroup)
 		return 0;
 	return mObjectGroup->getWidth();
 }
 
-int Extension::getObjectGroupHeight()
-{
+int Extension::getObjectGroupHeight() {
 	if (!mObjectGroup)
 		return 0;
 	return mObjectGroup->getHeight();
 }
 
-float Extension::getObjectGroupOpacity()
-{
+float Extension::getObjectGroupOpacity() {
 	if (!mObjectGroup)
 		return 1;
 	return mObjectGroup->getOpacity();
 }
 
-int Extension::isObjectGroupVisible()
-{
+int Extension::isObjectGroupVisible() {
 	if (!mObjectGroup)
 		return 1;
 	return mObjectGroup->getVisible();
 }
 
-char *Extension::getObjectGroupProperty(char *name, char *defaultValue)
-{
+char *Extension::getObjectGroupProperty(char *name, char *defaultValue) {
 	if (!mObjectGroup)
 		return __(defaultValue);
 	std::map<std::string, std::string> properties = mObjectGroup->getProperties();
@@ -360,64 +312,55 @@ char *Extension::getObjectGroupProperty(char *name, char *defaultValue)
 	return __(defaultValue);
 }
 
-char *Extension::getObjectName()
-{
+char *Extension::getObjectName() {
 	if (!mObject)
 		return __("");
 	return __(mObject->getName().c_str());
 }
 
-char *Extension::getObjectType()
-{
+char *Extension::getObjectType() {
 	if (!mObject)
 		return __("");
 	return __(mObject->getType().c_str());
 }
 
-int Extension::getObjectPosOnMapX()
-{
+int Extension::getObjectPosOnMapX() {
 	if (!mObject)
 		return 0;
 	return (int) mObject->getX() + mOffsetX;
 }
 
-int Extension::getObjectPosOnMapY()
-{
+int Extension::getObjectPosOnMapY() {
 	if (!mObject)
 		return 0;
 	return (int) mObject->getY() + mOffsetY;
 }
 
-int Extension::getObjectWidth()
-{
+float Extension::getObjectWidth() {
 	if (!mObject)
 		return 0;
-	return (int) mObject->getWidth();
+	return mObject->getWidth();
 }
 
-int Extension::getObjectHeight()
-{
+float Extension::getObjectHeight() {
 	if (!mObject)
 		return 0;
 	return mObject->getHeight();
 }
 
-float Extension::getObjectRotation()
-{
+float Extension::getObjectRotation() {
 	if (!mObject)
 		return 0;
 	return mObject->getRotation();
 }
 
-int Extension::getObjectTileGid()
-{
+int Extension::getObjectTileGid() {
 	if (!mObject)
 		return 0;
 	return mObject->getGid();
 }
 
-char *Extension::getObjectVertices(char *defaultValue)
-{
+char *Extension::getObjectVertices(char *defaultValue) {
 	if (!mObject)
 		return __(defaultValue);
 	const char *vertices = mObject->getVertices().c_str();
@@ -426,16 +369,14 @@ char *Extension::getObjectVertices(char *defaultValue)
 	return __(vertices);
 }
 
-char *Extension::getObjectBox2DVertices(char *defaultValue)
-{
+char *Extension::getObjectBox2DVertices(char *defaultValue) {
 	if (!mObject)
 		return __(defaultValue);
 	char *vertices = __(mObject->getVertices().c_str());
 	if (!vertices || !_strcmpi(vertices, __("")))
 		return __(defaultValue);
 	int i = 0;
-	while (vertices[i] != '\0')
-	{
+	while (vertices[i] != '\0') {
 		if (vertices[i] == ' ')
 			vertices[i] = ',';
 		++i;
@@ -443,57 +384,49 @@ char *Extension::getObjectBox2DVertices(char *defaultValue)
 	return __(vertices);
 }
 
-int Extension::isObjectPolygon()
-{
+int Extension::isObjectPolygon() {
 	if (!mObject)
 		return 0;
 	return mObject->getPolygonType() == "polygon";
 }
 
-int Extension::isObjectPolyline()
-{
+int Extension::isObjectPolyline() {
 	if (!mObject)
 		return 0;
 	return mObject->getPolygonType() == "polyline";
 }
 
-int Extension::isObjectEllipse()
-{
+int Extension::isObjectEllipse() {
 	if (!mObject)
 		return 0;
 	return mObject->getPolygonType() == "ellipse";
 }
 
-int Extension::isObjectFlippedHorizontally()
-{
+int Extension::isObjectFlippedHorizontally() {
 	if (!mObject)
 		return 0;
 	return mObject->getFlippedHorizontally();
 }
 
-int Extension::isObjectFlippedVertically()
-{
+int Extension::isObjectFlippedVertically() {
 	if (!mObject)
 		return 0;
 	return mObject->getFlippedVertically();
 }
 
-int Extension::isObjectFlippedDiagonally()
-{
+int Extension::isObjectFlippedDiagonally() {
 	if (!mObject)
 		return 0;
 	return mObject->getFlippedDiagonally();
 }
 
-int Extension::isObjectVisible()
-{
+int Extension::isObjectVisible() {
 	if (!mObject)
 		return 0;
 	return mObject->getVisible();
 }
 
-char *Extension::getObjectProperty(char *name, char *defaultValue)
-{
+char *Extension::getObjectProperty(char *name, char *defaultValue) {
 	if (!mObject)
 		return __(defaultValue);
 	std::map<std::string, std::string> properties = mObject->getProperties();
